update CI so that commit is more specific
Builds have been failing for about two weeks because an additional file was added to the setup miniconda step (I think) and that file was commited to the data branch. Subsequent checkouts of the data branch from main was causing problems. I have removed this file on the data branch and made it so only csv files will be added from now on.